SERVER-103191 Resharding coordinator handles missing state doc (#34396)
[SERVER-103191](https://jira.mongodb.org/browse/SERVER-103191) made the
coordinator read the latest version of its state document before
reporting its completion log line to ensure that it had the latest
changes on disk written by the participants. However, this read occurs
within the same retry loop as removing the state document at the end of
the operation, so it's possible that the coordinator deletes the state
document, receives some transient error, retries, then attempts to read
the (now missing) state document, ultimately leading to a fatal error.

This change makes the coordinator attempt to read the state document if
it exists, otherwise it uses its in-memory copy of the state document.
If a retry occurs, this should contain the latest version of the state
document before it was deleted.