Allow descriptors to be stateful in SDK (#34473)
Stateful descriptors allows us to easily curry state that is shared
between stages (threads, connections) and use
arbitrary code to assemble this state through extension initialization.

Factor out tokio runtime creation to happen in or near extension
initialization.

Make voyage rerank registration conditional on the presence of the api
key in the env. Not sure if this is preferred,
an easy alternative would be to fail bound descriptor creation if the
key is not present.

---------

Co-authored-by: Will Buerger <59492746+wbuerger46@users.noreply.github.com>