Make source reads in transform stages exception safe (#34888)
Catch exceptions when reading from the source and pass them to the
extension caller, who may then pass them
back to the host where they are re-thrown. This arrangement ensures that
we never unwind through an extension
stack, even though this would be safe with our rust SDK due to use of
the `C-unwind` ABI.

To do this we introduce a `MongoExtensionError` type that for now is
only used by executor::getNext(). When an
exception occurs invoking a host executor, it is captured, wrapped in
this type, and passed to the extension.
`DocumentSourceExtension` examines error when pulling from the stream
and if the vtable matches a host thrown
exception then it will re-throw the exception.

The rust SDK accommodates and wraps external errors to allow executor
implementations to simply pass them through.
This requires some relatively small API changes in error handling.

I've flagged some of the tricky bits in the code, please ask me lots of
questions about anything you don't understand.