Defer tokio runtime creation to first use (#35166)
This fixes an issue where the server wasn't shutting down cleanly.

The cause of the unclean shutdown was that the wrong thread(s) were
receiving SIGTERM and using a default
handler that terminates the process immediately instead of going through
database shutdown. The ultimate cause
of this is that tokio runtimes/threads were being created at stage
registration during global initializers, which happens
before the signal processing thread is created and the main thread
signal mask is adjusted, so tokio threads were
inheriting an incorrect signal mask.

In the long run we will likely load extension dynamically _after_ the
signal processing thread is created, so that
any threads created by extensions will inherit the correct signal mask.

h/t to @sroches for locating where sigmask is adjusted to the correct
values.