[PoC] Implement $betaMultiStream (#35134)
This adds a draft of $betaMultiStream stage, which can be used for one
pipeline to handle two streams of data.

Right now, there are 2 "methods" with which to handle the secondary
stream: a "cursor" method, which populates a secondary cursor with the
results, and a "setVar" method, which populates the $$SEARCH_META
variable with the results. (If implemented outside of this spike, there
are many opportunities to improve the syntax and expose more
functionality, like setting an arbitrary variable name.)

The intended use case is for a stage like `$search` to desugar to
something like
```
{$betaMultiStream: {
  primary: [{$_internalPluginSearch}, {$_internalIdLookup}],
  secondary: [{$_internalPluginSearchMeta}],
  method: "setVar"
}}
```

On a single node, that stage will be transformed into an executor that
looks like
```
[
  {$_internalPluginSearch},
  {$_internalIdLookup},
  {$setVariableFromSubpipeline: {var: "$$SEARCH_META", pipeline:[{$_internalPluginSearchMeta}]}},
  ... remaining stages
]
```

In a sharded cluster, the router will split the pipeline into:
```
// Shards pipeline; each shard returns 2 cursors.
[
{$betaMultiStream: {
  primary: [{$_internalPluginSearch}, {$_internalIdLookup}],
  secondary: [{$_internalPluginSearchMeta}],
  method: "cursor"
}}
]

// Merging pipeline.
[
  {$sort: <$search sort spec>},
  // The second cursor from each shard will feed into this subpipeline.
  {$setVariableFromSubpipeline: {var: "$$SEARCH_META", pipeline:[{$_internalPluginSearchMeta}]}},
  ... remaining stages
]
```

## Notably Left Out
Right now, $betaMultiStream can only be used with stages that source
data themselves, not stages that open a cursor to read from WiredTiger.
This was to, for now, avoid the complexity of having two pipelines share
a source.

_There are many open design questions for this beta stage that should be
discussed in more depth if a stage like this is to be implemented on
master._