Fix rust build script missing header errors (#36363)
Some rust crates compile C or C++ dependencies, using the cc toolchain
configured in the environment to do so.
As part of this it pulls toolchain compiler flag settings and passes
them to the `cc` crate. Bazel cc toolchains are typically
hermetic and use `-no-canonical-prefixes` so the compiler won't auto
detect header locations and are instead redirected
to a sysroot. Rust build scripts are run a bit like a genrule and have
the toolchain symlinked into the build directory so the
related macros make `-isystem` paths absolute. The mongo toolchain
writes this as `-isystemrelative/path` which
`rules_rust` does not know how to process. Fix the mongo `-isystem`
includes.

Add a dependency on the `simsimd` crate which fails before this change
because it can't find `arm_neon.h`.

This is necessary but not sufficient to fix `opt` builds. Among the
problems:
1. The version of gcc 14 the mongo toolchain uses can't compile clang
for bindgen
2. g++ fails to link proc-macro crates, which is weird because those
crates probably shouldn't be linked at all.